requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
anthropic>=0.40.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
